Add business rules for the organization and contact tables
Add BusinessRulesUtils script includes to group logic used for business rules
that is cannot be done in  the regular business rule conditions (i.e., no script
being used). The match regex condition does not appear to be working
properly, so this is done in a script.

If an error is caught by the business rule, rather than aborting the insert or
update action, the field that has the error is set to null to prevent the lost
of other data being inserted/updated in the record.

Ticket: AT-7107