Add table and flow for Salesforce COR/ACOR updates
Add a basic table with a 'request body' column having
a max length of 4000 for sending a request. The 'sys_id'
auto generated by SNOW is considered to be the 'id'.

Add a flow that sends an email to atat-dev@ccpo.mil
when a new record is created in the above table. This
email request appears to be for updating COR/ACOR
information in Salesforce and the email in the flow
will be changed.

Ticket: AT-7102