Add military rank table to break up rank components
Add military rank table to provide different components
that relate to the rank when a contact role is 'Military'.
- branch - choice
- rank name - string
- rank abbr - string
- grade - choice

Update 2 contact choices to include (FLTCM) and (FORCM) 
which allows the rank choices to be parsed with a script.

Create military rank records by creating a script to parse 
the label field of the 'rank' choices, then  insert the records 
into the military table.

Ticket: AT-7144