Add an action and flow to provision jobs in AWS
Create an action (AWS Provisioning Job Request) that
takes in parameters to be used for making the
provisioning request body. Some of these are hard
coded since there is no place in ATAT to get the data.
No credentials are currently needed to make the
call to AWS.

Add a flow (ATAT Provisioning Request Flow) to create
a provisioning job record and start the provisioning
process in AWS using the AWS Provisioning Job Request
action.

Ticket: AT-7078