Create a getProvisioningJobResults function
Implement a script includes function to call from a global scheduled
job. The function sends an outbound REST message request
(GET /provisioning-jobs to AWS) to poll messages on the provisioning
job queue.

The response is looped through to get the 'cspResponse' in each
message and update the provisioning job record in SNOW. A new
portfolio is created (in SNOW) if the provisioning job is successful.

Ticket: AT-7057