Revert order of subflows
Revert subflows order by placing the DocuSign subflow first.
When the generate DoW subflow is first the DoW is created, 
however, the DoW document is not attached to the acquisition
package.

There does not appear to be a clear cause since the DoW document is
created, but is simply not attached when the generate DoW subflow is 
first. 

Ticket: AT-7482