Update flow to run subflows in parallel
Update ATAT Prepare Acquisition Package Documents flow
to run each document generation/creation subflow in parallel. 

The generate DoW subflow was not being attached to the acq. package
and as a solution a SNOW community post suggested adding a 5 second
wait prior to running the subflow, which solved the issue.

Ticket: AT-7482