Create docs folder at top level
The README.md file inside of the f600233d1b154d507b782f84604bcb12
folder continues to remove the end of the file. To prevent this
it has been moved to the docs/ folder at the top level, since
generated files are found in f600233d1b154d507b782f84604bcb12 and
is considered less reliable.