Add script include and REST API for manually creating multipart body (WIP)
Add the below items to be able to manually generate a multipart body that is stored
in a temp attachment, which can then be used as the request body.
This provides more control over the multipart body being sent in a request to
third party systems. This is a rough make shift version to try and get it to work first.
- EDMS (REST API)
- GET /edms-multipart-attachment resource
- MultipartHelper script includes (calls the EDMS resource to create an attachment with
using RESTAPIResponse object getStreamWrite(), which is only available in REST API resource
scripts)

Inspiration from links below
- https://blog.tgxn.net/multipart-upload-from-servicenow/
- https://developer.servicenow.com/connect.do#!/share/contents/8774514_multipart_upload_helper?t=PRODUCT_DETAILS