Add script include and REST API for manually creating multipart body (WIP)
This commit is a squashed commit for the high-code solution for uploading documents
to EDMS.

Add the below items to be able to manually generate a multipart body that is stored
in a temp attachment, which can then be used as the request body.
This provides more control over the multipart body being sent in a request to
third party systems. This is a rough make shift version to try and get it to work first.
- EDMS (REST API)
- GET /edms-multipart-attachment resource
- MultipartHelper script includes (calls the EDMS resource to create an attachment with
using RESTAPIResponse object getStreamWrite(), which is only available in REST API resource
scripts)

Inspiration from links below
- https://blog.tgxn.net/multipart-upload-from-servicenow/
- https://developer.servicenow.com/connect.do#!/share/contents/8774514_multipart_upload_helper?t=PRODUCT_DETAILS

-------
Add user properties for user for EDMS API Resource

- Add system properties for 'ATAT Client'

-------
Update edms-multipart-attachment (REST script)

- Update boundary in parts
- Update content-type for file to application/octet-stream
- Add additional header in the file section for base64 (commented out)
- Add section that sends base64 content from test.pdf (commented out)

-------
Fix multipart request construction

- Remove "content-type" from multipart
- Add an additional "\r\n\" to "content-disposition"

// Tried as a fix but does not appear to be the cause of error
- format indexFields as a raw string  (commented out)

-------
Success with request

- Create All Levels set flag to true added for the post-award

Ticket: AT-7754