Update EDMS: Upload Document action
Add a script step that contains logic for manually constructing the multipart
request for uploading a document to EDMS (aka high-code solution).

Remove the previous implementation that was focused on using the REST
step (aka low-code solution). There was a bug that prevented the use
of the RESt step, and the bug fix is planned to be fixed in the Utah release
(current release is Rome).