Fix issues with setting flow variables outside of for loop
Solution:
- remove the action to set the flow variables to zero above the flow loop
- initialize flow variables based on a ternary operation inside the loop

Issue related to adding flow variables for successful/failed document uploads.
When trying to initialize the count to zero for successful/failed uploads flow
variable the flow did not appear to run properly and would throw an unknown
error which seemed related to a null pointer and viewing the run results was not
possible.

It is still unclear why this is causing an error since similar structures have been used
in previously subflows that set an initial value for flow variables  for use later in the
subflow.