Create 'Copy Package Attachments to Email' action
- Create 'Copy Package Attachments to Email' action to attach gathered
documents to the email to send to DITCO using the 'EmailAttachments'
script include and GlideSysAttachment.copy().

Using the GlideSysAttachment.copy() inside of the 'EmailAttachments'
script include did not properly copy when running the flow, so the copy()
method was moved into the custom script of the action, which works as
expected.