Update DOCGEN: Send Package to DITCO flow
- Add 'Copy Attachments to Email' action
- Look up funding request attachments
- Loop over found attachments and attach to email using 'Copy Attachment' action

The Funding Request attachments are the only documents copied using the core flow 
designer 'Copy Attachment' action. This was done because the funding request attachments
are stored in a file attachment column preventing the GlideSysAttachment.copy() method 
from working because of the prefix "ZZ_YY". The 'Copy Attachment' action  was used as a 
work around, but could possibly be used to refactor and copy the final documents in the 
Signed/Unsigned records.