Refactor 'PackageAttachments' script include
Refactor 'PackageAttachments' to account for documents being gathered
after document generation, due to some logic occurring prior to sending an email
to DITCO (e.g., submitting the package). Therefore, documents are gathered prior
to sending an email for users to download documents and includes an intermediate
state where the user is uploading unsigned and signed documents.

- Rename method 'gatherUnsignedDocumentsToEmail' to 'prepareUnsignedDocumentsToEmail'
because the documents are gathered in a an earlier step and there is not much required other than
cleaning up documents not needed at the moment

- Refactor 'deleteSignedDocumentAttachments' method to be more flexible to delete documents
based on an array of file names, and changing the name to 'deleteUnsignedDocumentAttachments' 

- Rename 'deleteUnsignedRecordAttachments' to 'deleteAllUnsignedDocumentAttachments' for
consistency