Update PackageAttachments script include to create unsigned/signed records
- Add logic to create an unsigned and signed  record for a package if the
records do not already exist.

It was considered to place the creation of the records in the
'DOCGEN: Prepare Acquisition Package Documents' flow or in the
'UTIL: Copy Package Attachments for Download' action. However,
most of the logic interacting with the two tables was provided in the
initialize method of the PackageAttachments script include, which was
already looking up the records if they exist.

The other locations would require adding what seemed like repeated logic to look
up the records and then create records if no records exists.